requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
PyQt5==5.15.6
faiss-cpu==1.7.1
sentence-transformers==2.1.0
scikit-learn==1.0.1
detectron2==0.6
yolov5==6.0.0
albumentations==1.1.0
segmentation-models-pytorch==0.2.1
mss==6.1.0
pyautogui==0.9.53
openai==0.27.0
learn2learn==0.1.7
Pillow
screeninfo